Improvements to the self-updater, etc.
There is a lot in this patch. Most of it is related to self-
installation and self-update.

It moves all the self-update code to multirust_cli::self_update, adds
tests to tests/cli-self-update.rs, and makes everything real solid.

Installation is hypothetically simple, as multirust is distributed as
a single binary, and installation mostly requires copying it into
place. There are some tricky bits though, mostly because of
workarounds to self-delete an exe on Windows.

During install (as `multirust-setup`):

* copy the self exe to $CARGO_HOME/bin
* hardlink rustc, etc to *that*
* update the PATH in a system-specific way
* run the equivalent of `multirust default stable`

During upgrade (`multirust self upgrade`):

* download multirust-setup to $CARGO_HOME/bin/multirust-setup
* run multirust-setup with appropriate flags to indicate
  this is a self-upgrade
* multirust-setup copies bins and handlinks into place. On windows
  this happens *after* the upgrade command exits successfully.

During uninstall (`multirust self uninstall`):

* Delete `$MULTIRUST_HOME`.
* Delete everything in `$CARGO_HOME`, including
  the multirust binary and its hardlinks

Deleting the running binary during uninstall is tricky
and racy on Windows.

The last step of uninstallation is to delete *the running binary*,
multirust.exe and the CARGO_HOME that contains it. On Unix, this
works fine. On Windows you can't delete files while they are open,
like when they are running.

Here's what we do:
- Copy multirust to a temporary file in
  CARGO_HOME/../multirust-gc-$random.exe.
- Open the gc exe with the FILE_FLAG_DELETE_ON_CLOSE and
  FILE_SHARE_DELETE flags. This is going to be the last
  file to remove, and the OS is going to do it for us.
  This file is opened as inheritable so that subsequent
  processes created with the option to inherit handles
  will also keep them open.
- Run the gc exe, which waits for the original multirust
  process to close, then deletes CARGO_HOME. This process
  has inherited a FILE_FLAG_DELETE_ON_CLOSE handle to itself.
- Finally, spawn yet another system binary with the inherit handles
  flag, so *it* inherits the FILE_FLAG_DELETE_ON_CLOSE handle to
  the gc exe. If the gc exe exits before the system exe then at
  last it will be deleted when the handle closes.

This is the DELETE_ON_CLOSE method from
http://www.catch22.net/tuts/self-deleting-executables

... which doesn't actually work because Windows won't really
delete a FILE_FLAG_DELETE_ON_CLOSE process when it exits.

.. augmented with this SO answer
http://stackoverflow.com/questions/10319526/understanding-a-self-deleting-program-in-c

The `.multirust` dir no longer lives in `AppData\Local` on Windows.
While that may be a better place for it, this change makes
multirust's behavior more consistent with Cargo's.

To transition from bins being in `.multirust/bin` to `.cargo/bin`,
multirust deletes those in the former on initial upgrade, when invoked
as `multirust-rs`. This includes those from the old `AppData`
location. It does *not* try to clean up the previous `PATH`
modifications, nor does it delete the `.multirust` directory.

The `multirust self install` command doesn't exist as that
functionality is entirely in `multirust-setup`.