Deploy to s3. Add simple website.
Name the setup program `rustup-setup.exe`. Modify travis and appveyor
configs to deploy it from master to dev-static.rust-lang.org, and from
the stable branch to static.rust-lang.org.

The directory structure under s3 is

rustup/
  www/ - the (simple) website
  rustup-setup.sh - the curlable script
  dist/$target/
    rustup-setup$exe
    rustup-setup$exe.sha256

The shell script and website are deployed from x86_64-linux.

Self upgrades are directed to

> https://static.rust-lang.org/rustup/dist/$target/

The rustup-setup.sh script is accessible via sh.rustup.rs,
one of the windows bins at win.rustup.rs, and the website
at www.rustup.rs.

The website is really simple. It does platform detection
and says how to use sh.rustup.rs or win.rustup.rs.

The website is not necessarily final, but I'm probably
not going to think much more about it before asking
for testing and feedback.

Once this is deployed I will do a one-time deployment of the
bins to the old upgrade location to provide continuity.