Add `rustup show`
This shows the active toolchain and the reason for selecting it.

In rbenv this is `rbenv version`, but we use the word 'toolchain'
in rustup. Unfortunately the `rustup toolchain` namespace is
taken with functions that specifically deal with toolchains,
whereas this command deals with toolchains, overrides and defaults.
That's why this isn't `rustup toolchain`.