Put the target triple at the end of toolchain names
$channel-$date-$triple instead of $triple-$channel-$date. This
emphasizes that the channel is the most important information
here (it's the only thing that's required). Putting the
component of the name with potentially unbounded values at the end
makes parsing potentially saner over the long term, makes it
possible in the future to parse toolchains without hard-coding
the exact architectures supported.