Mentioned Visual C++ Build Tools 2015
There is a Visual C++ Build Tools 2015 that installs just the C++ build tools without the full Visual Studio. It includes the necessary linker to build Rust MSVC. It's on the same downloads page, under "Tools for Visual Studio 2015". I think a lot of people would be happy to know of a more lightweight installer option.
More info here: https://blogs.msdn.microsoft.com/vcblog/2016/03/31/announcing-the-official-release-of-the-visual-c-build-tools-2015/