symlink the rustup-init binary during tests
We can't open it for writing as any concurrent forks would keep a writable file
descriptor open which would prevent us from then executing the file with a "text
file busy" error.

On windows we keep copying as this problem isn't present and symlinks don't work
that well on Windows.