Add 'default'/'update' bare triple handling with erroring out
When a bare (partial) triple is given instead of a toolchain name, try to
figure out user's intention. For 'update', generate a list of toolchains
with the matching triple. For 'default', find out if there's a non-default
toolchain on the same channel/date and the given triple, and suggest its
use. In both cases, terminate with an error, so the user has to reissue
the command with a valid toolchain spec.