Update tests for the rustup-dist package to pass around DownloadCfg
Also replaced the usage of download::DownloadCfg with a simple call
to utils::download_file, to simplify the number of names being used
in the file.

The error behaviour when a component fails to download has changed
slightly; underlying errors get wrapped in a ComponentDownloadFailed,
rather than propogating out of manifest::update directly. I think this
should be okay because the error_chain doesn't throw the info away.