Add the rust lib dir (containing std-<hash>.dll) to the path on windows
Adds sysroot/bin to the path on Windows which fixes finding libstd in
dynamic builds. This fixes a load of dylib tests in the cargo testsuite
on Windows. This will result in rustc.exe being directly in the path
although it is after CARGO_HOME/bin so the shim should be found first.