Prevent `rustup update` to a toolchain without `rustc` or `cargo`.
When the Rust CI is misconfigured, it is possible to produce a nightly
where `rustc` or `cargo` is not available. This makes the entire toolchain
unusable (for one day at least).

This PR tries to prevent the update operation from starting at all if
these essential components are missing, so the user would still have a
working though outdated toolchain.

Resolves #1297.