CI: Dump testament during build
To allow us to ensure that we do not have unexpected version strings
due to dirty builds etc, dump the testament after the build but before
running the tests.  We can't do this for unusual cross-builds but
for the most part we can do it on at least one of each of our major
platforms.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>