Show the time spent doing downloads
It can be useful to know how long it takes to do a download, and
similarly while we use this to track installs, to do installs.

Looking at this I identified that if no data is received for a second
the current code overestimates the download performance (the missing)
seconds are elided and not counted towards averages. I've made a note
but fixing it is a different, larger change.