Fix #1870: Race condition with some virus scanners
Some virus scanners take out handles on files we are going to
rename, causing access errors and preventing the rename. Typically
this is short lived. Retry rather than erroring.

No feedback given at the moment, and it will be a much larger patch to
do so - as this is a regression fix I'd like to get it in and released
and follow up with something thread all the UI channels down to this
layer later (e.g. an implicit context mechanism of some sort perhaps?)