self-update: Detect when running under snapd or others
When we're running under snapd we don't want to self-update.
However we want to do that nice and quietly.  On the other hand
if we're explicitly self-updating under snapd we need to say no
to the user usefully.

In addition, if we can't write to the directory rustup was within
then we're likely running under a package manager of some kind.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>