Toolchain: Ensure we only list available components.
Extensions might have a target other than that of the toolchain
so this change ensures we use the extension's target not the target
of the toolchain when determining if it's available.  This will as
a result prevent `rustup target list` listing known but unavailable
targets if `rust-std` is unavailable for that target.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>