Make list_components safer to work with
PR 2116 introduced a regression in linked toolchain support. The
regression was introduced in part because there was no type information
signifying the difference between uninstalled toolchains and toolchains
that don't have components, thus the compiler didn't hint during
development that any discrimination was needed.

This patch introduces a type level discrimination to avoid future issues
of this type. The code is a little crufty due to error_chain glue
issues, but there is a longer term path to making that clean.