Avoid parsing TOML manually
When trying to detect the format of the toolchain override file,
let serde do all the work. If the file does not parse as TOML,
fallback to the legacy format, but keep the TOML error. If the first
line does not happen to be a valid toolchain, show also the TOML error
so that TOML syntax errors are reported.