Don't access the Internet in the check_updates tests
Previously, three of the tests in `cli-exact.rs` would try to connect to the
Internet, or fail if the Internet wasn't available:

*   `check_updates_with_update`
*   `check_updates_none`
*   `check_updates_some`

This is because all three tests would run `rustup check`, and because the tests
weren't overriding the value of `RUSTUP_UPDATE_ROOT`, they'd go out to the
Internet to try to download self-updates.

The new helper mimics the behaviour of the `self_update_setup()` helper in the
same file.  It creates a mock self-update server in a local directory, and
points rustup to the directory using a `file://` URL.  When it tries to get
self-updates as part of `rustup check`, it only fetches from that directory.