override: Detect v1 manifests and do not reinstall if found
When we're installing a toolchain via the override mechanism
we try and ensure components requested are present.  This is
a no-op on v1 manifest installations but we still end up downloading
and checking the update hash.  This is unnecessary, so this change
attempts to prevent this by detecting a V1 install and skipping
the component check if so.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>