Optimization: avoid double-parsing the manifest
This adds an intermediate struct which includes a parsed manifest, and
eliminates a bunch of duplicated code.  Only parsing the manifest once
speeds up `find_or_install_override_toolchain_or_default` by about 50%
which is the critical path for a rebuild with no changes.

There's also a drive-by fix for running test on macOS aarch64.