Remove snapcraft build and runtime support entirely
Snap was a best effort effort for Rustup, is currently broken, and we
have no capacity to develop it.

Remove it to signal clearly that issues with it are not something we're
focusing on - allowing other people to step up and maintain a
distribution of Rustup for Snap if desired.