Bump remove_dir_all
While this fixes a TOCTOU bug in the dependency, rustup is run either as
an unprivileged process, or as a privileged process operating on a file
tree that the privileged user owns; it is not a setuid binary and
that mitigates much of the vulnerability: a deliberately misconfigured
sudoers or similar would be required. Thus this hardens rustup against
misconfiguration, but does not close an active vulnerability in how
rustup is designed to be used.